refactor(userspace-dp): extract afxdp/ha state ownership
Move the HA state-transition ownership group out of afxdp/ha/mod.rs into
a dedicated `state` submodule (#6247). Pure code motion: method bodies are
copied verbatim; only the module glob (`use crate::afxdp::*`) and the impl
target path are rewritten, both semantically identical for a descendant of
`afxdp`.

Relocated inherent methods:
- update_ha_state (RG-epoch publication ordered before the ArcSwap runtime
  publish, demote/activate fan-out, cache-flush timestamp)
- handle_activated_rgs (activation prewarm + reverse republish + neighbor
  warm)
- ha_groups (status projection)
- last_cache_flush_at (observability accessor)

No behavior, signature, ordering, or visibility change. Release RG-epoch
publish-before-runtime-visibility and post-publish command ordering are
preserved byte-for-byte.
